Release: Player Companies, Company Creation Tax & Fixes (#34)
* Feature: Basic GUIs to buy and sell stocks. (#18)

* feat: added some simple guis to manage portfolio and companies.

* fix: several graphic issues with the new guis.

* Fix: removed static plugin instance. (#19)

* Fix: Update to paths in order to fix windows bootstrap (#20)

* Fix: Format high numbers with a more human-readable format (#22)

* Feature: Button to buy 100 Company shares in Companies GUI. (#23)

* feat: action to buy 100 stocks of a company with one click.

* fix: fixed company variation percentage in signs.

* Feature: Player-oriented Company Creation and Taxes (#29)

* feat: commands for players to create/delete companies.

* feat: configurable taxes per company risk.

* Bugfix: Companies created with past IDs (#30)

* Bugfix: Players can control admin-created companies (#31)

* Bugfix: Fixing argument parsing for Risk and Share price (#32)

* chore: removed unnecessary messages.

* chore: update plugin version to 2.5.0

* Bugfix: Investment item showing negative values in GUI (#33)

+++ src/main/java/dev/hugog/minecraft/blockstreet/commands/DeleteCommand.java
@@ -2,46 +2,56 @@ package dev.hugog.minecraft.blockstreet.commands;
 
 import dev.hugog.minecraft.blockstreet.data.dao.CompanyDao;
 import dev.hugog.minecraft.blockstreet.data.services.CompaniesService;
+import dev.hugog.minecraft.blockstreet.data.services.PlayersService;
 import dev.hugog.minecraft.blockstreet.utils.Messages;
 import dev.hugog.minecraft.dev_command.annotations.*;
 import dev.hugog.minecraft.dev_command.arguments.parsers.IntegerArgumentParser;
 import dev.hugog.minecraft.dev_command.commands.BukkitDevCommand;
 import dev.hugog.minecraft.dev_command.commands.data.BukkitCommandData;
+import net.milkbowl.vault.economy.Economy;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 
 import java.text.MessageFormat;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
 /**
  * Delete Company Command
  *
- * <p>Command that allow players to delete a company.
+ * <p>Command that allows players to delete a company.
  * <p>Syntax: /invest delete [ID]
  *
  * @author Hugo1307
  * @since v1.0.0
  */
 @AutoValidation
-@Command(alias = "delete", description = "Delete a company.", permission = "blockstreet.command.delete")
+@Command(alias = "company delete", description = "Allow players to delete their company.", permission = "blockstreet.command.delete", isPlayerOnly = true)
 @Dependencies(dependencies = {Messages.class, CompaniesService.class})
 @Arguments({
-        @Argument(name = "companyId", description = "The ID of the company to delete.", position = 0, parser = IntegerArgumentParser.class)
+        @Argument(name = "id", description = "The id of the company to delete.", position = 0, parser = IntegerArgumentParser.class)
 })
 public class DeleteCommand extends BukkitDevCommand {
 
+    private final Messages messages;
     private final CompaniesService companiesService;
+    private final PlayersService playersService;
+    private final Economy vaultEconomy;
 
     public DeleteCommand(BukkitCommandData commandData, CommandSender commandSender, String[] args) {
         super(commandData, commandSender, args);
+        this.messages = getDependency(Messages.class);
         this.companiesService = getDependency(CompaniesService.class);
+        this.playersService = getDependency(PlayersService.class);
+        this.vaultEconomy = getDependency(Economy.class);
     }
 
     @Override
     public void execute() {
 
-        Messages messages = getDependency(Messages.class);
-        long companyId = Long.parseLong(getArgs()[0]);
+        Player player = (Player) getCommandSender();
+        long companyId = (int) getArgumentParser(0).parse().orElseThrow();
 
         if (!companiesService.companyExists(companyId)) {
             getCommandSender().sendMessage(messages.getPluginPrefix() + messages.getInvalidCompany());
@@ -49,20 +59,31 @@ public class DeleteCommand extends BukkitDevCommand {
         }
 
         CompanyDao companyToDelete = companiesService.getCompanyById(companyId);
+        if (!playersService.hasSharesInCompany(player.getUniqueId(), companyToDelete.getId(), companyToDelete.getTotalShares())) {
+            getCommandSender().sendMessage(messages.getPluginPrefix() + messages.getCannotDeleteNotOwnedCompany());
+            return;
+        }
+
+        double companyValuation = companyToDelete.getTotalShares() * companyToDelete.getCurrentSharePrice();
+
+        playersService.removeSharesFromPlayer(player.getUniqueId(), companyToDelete.getId(), companyToDelete.getTotalShares());
+        vaultEconomy.depositPlayer(player, companyValuation);
         companiesService.deleteCompany(companyId);
 
+        getCommandSender().sendMessage(messages.getPluginPrefix() + MessageFormat.format(messages.getSoldActions(), companyToDelete.getTotalShares(), companyValuation));
         getCommandSender().sendMessage(messages.getPluginPrefix() + MessageFormat.format(messages.getDeletedCompany(), companyToDelete.getName()));
 
     }
 
     @Override
     public List<String> onTabComplete(String[] args) {
+        Player player = (Player) getCommandSender();
         if (args.length == 1) {
-            return companiesService.getAllCompanies().stream()
-                    .map(CompanyDao::getId)
-                    .map(String::valueOf)
+            return playersService.getInvestments(player.getUniqueId()).stream()
+                    .filter(investment -> companiesService.companyExists(investment.getCompanyId()) && companiesService.getCompanyById(investment.getCompanyId()).getTotalShares() <= investment.getSharesAmount())
+                    .map(investment -> String.valueOf(investment.getCompanyId()))
                     .collect(Collectors.toList());
         }
-        return List.of();
+        return Collections.emptyList();
     }
 }
